+++ options.js
@@ -11,12 +11,29 @@ document.getElementById('help').addEventListener('click', function() {
 // Saves options to chrome.storage
 const saveOptions = () => {
     const type = document.getElementById('type').value;
+    const defspace = document.getElementById('defspace').value;
     const spacekey = document.getElementById('spacekey').value.toUpperCase();
-    const subdomain = document.getElementById('subdomain').value;
+    var rooturl = document.getElementById('rooturl').value;
     const limit = document.getElementById('limit').value;
+    const advancedsearch = document.getElementById('advancedsearch').checked;
   
+    // fix rooturl
+    // remove trailing slash if present
+    rooturl = rooturl.replace(/\/$/, "");
+    // append .atlassian.net if only subdomain is provided (without .)
+
+    // ensure rooturl starts with http or https
+    if (!rooturl.match(/^http/)) {
+        rooturl = `https://${rooturl}`;
+    }
+
+    // append /wiki if url ends with atlassian.net
+    if (rooturl.match(/\.atlassian\.net$/)) {
+        rooturl += '/wiki';
+    }
+    
     chrome.storage.sync.set(
-      { type: type, spacekey: spacekey, subdomain: subdomain, limit: limit },
+      { type: type, spacekey: spacekey, defspace: defspace, rooturl: rooturl, limit: limit, advancedsearch:advancedsearch },
       () => {
         // Update status to let user know options were saved.
         const status = document.getElementById('status_msg');
@@ -26,18 +43,40 @@ const saveOptions = () => {
         }, 750);
       }
     );
+
+    // Request permission for the domain only for url not hosted on atlassian.net
+    if (!rooturl.includes('.atlassian.net')) {
+      chrome.permissions.request(
+        {
+          origins: [rooturl + '/'] // Append trailing slash for permission request else error Invalid value for origin pattern ... Empty path.
+          // If you also need API permissions:
+          // permissions: ["tabs"]
+        },
+        function(granted) {
+          if (granted) {
+            // Permission was granted
+            console.log(`Permission granted for ${rooturl}!`);
+          } else {
+            // Permission was denied
+            console.log(`Permission denied for ${rooturl}!`);
+          }
+        }
+      );
+    }; // end if not atlassian.net
   };
   
   // Restores select box and checkbox state using the preferences
   // stored in chrome.storage.
   const restoreOptions = () => {
     chrome.storage.sync.get(
-      {type: 'page', spacekey: '',subdomain:'', limit:25 },
+      {type: 'page',defspace: 's', spacekey: '',rooturl:'', limit:25, advancedsearch: false },
       (items) => {
         document.getElementById('type').value = items.type;
+        document.getElementById('defspace').value = items.defspace;
         document.getElementById('spacekey').value = items.spacekey;
-        document.getElementById('subdomain').value = items.subdomain;
+        document.getElementById('rooturl').value = items.rooturl;
         document.getElementById('limit').value = items.limit;
+        document.getElementById('advancedsearch').checked = items.advancedsearch;
       }
     );
   };